Rewrite Fields Module
The Fields module was the one that needed the most work. It was the first one I coded and the one that I used Winterbraid's code as a reference for. Now, it has been rewritten from the ground up to be as tight and efficient as possible. Unless the site's HTML changes, this should be the final version of the code.